update_state_frame to use boxing
update_state_frame to use boxing. If i understand the current
implementation correctly then my change will work. What is happening is
we are deciding that the first box on the value list is what we want to
update. We are getting that box using the firstValueFrame abstraction
then passing that box and the new value to set-box! to update the value